Made API RESTful with put, post, patch, delete, and get being used in the appropriate methods.
