+++ ToDoFunctions/DeleteToDoItem.cs
@@ -13,11 +13,11 @@ namespace ToDoFunctions
     public static class DeleteToDoItem
     {
         [FunctionName("DeleteToDoItem")]
-        public static async Task<HttpResponseMessage> Run([HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "Api/DeleteToDoItem")]HttpRequestMessage req, [Table("todotable", Connection = "MyTable")]CloudTable table, TraceWriter log)
+        public static async Task<HttpResponseMessage> Run([HttpTrigger(AuthorizationLevel.Anonymous, "delete", Route = "Api/ToDoItem/{id}")]HttpRequestMessage req, string id, [Table("todotable", Connection = "MyTable")]CloudTable table, TraceWriter log)
         {
-            var val = req.Content;
-            var id = val.ReadAsStringAsync().Result;
-            id = id.Replace("\"", "");
+            //var val = req.Content;
+            //var id = val.ReadAsStringAsync().Result;
+            //id = id.Replace("\"", "");
 
             var item = Utility.GetToDoItemFromTable(table, id);
