+++ ToDoFunctions/Pages/Create.html
@@ -36,11 +36,11 @@
         var json = JSON.stringify(form);
         var xhr = new XMLHttpRequest();
 
-        xhr.open("POST", "/Api/AddToDoItem", true);
+        xhr.open("POST", "/Api/ToDoItem", true);
         xhr.setRequestHeader('Content-Type', 'application/json');
         xhr.onreadystatechange = function () {
             if (xhr.status == 201) {
-                window.location.href = "/Page/Index";
+                window.location.href = "/";
             }
         };
         xhr.send(json);
