+++ ToDoFunctions/Pages/Update.html
@@ -27,16 +27,26 @@
 <script>
 window.onload = function () {
     var url = window.location.href;
-    var editId = JSON.stringify(url.split("/Update/")[1]);
+    var editId = url.split("/Update/")[1];
+    var xhr = new XMLHttpRequest();
 
-    $.get("/Api/UpdateToDoItem", {id: editId}, function (data) {
-        $("#Title").val(data.Title);
-        $("#Description").val(data.Description);
-        $("#RowKey").val(data.RowKey);
-        var correctDate = data.Due.split("T");
+    xhr.open("GET", "/Api/ToDoItem/" + editId, true);
+    xhr.setRequestHeader("Content-Type", "application/json");
+    xhr.onreadystatechange = function () {
+        if (xhr.status == 200) {
+            var json = xhr.responseText;
 
-        $("#Due").val(correctDate[0]);
-    });
+            var data = JSON.parse(json);
+            $("#Title").val(data.Title);
+
+            $("#Description").val(data.Description);
+            $("#RowKey").val(data.RowKey);
+            var correctDate = data.Due.split("T");
+
+            $("#Due").val(correctDate[0]);
+        }
+    };
+    xhr.send();
 };
 
 $("#todoform").submit(function (event) {
@@ -49,9 +59,9 @@ $("#todoform").submit(function (event) {
         RowKey: $("#RowKey").val()
     };
     var json = JSON.stringify(form);
-    var xhr = new XMLHttpRequest();
+    xhr = new XMLHttpRequest();
 
-    xhr.open("POST", "/Api/UpdateToDoItem", true);
+    xhr.open("PATCH", "/Api/ToDoItem", true);
     xhr.setRequestHeader("Content-Type", "application/json");
     xhr.onreadystatechange = function () {
         if (xhr.status == 201) {
